remove warnings from console

+++ inc/mtb/mtb_chi_all_posts.php
@@ -90,6 +90,8 @@ function chi_selected_articles_or_video_save()
     // 	return;
     // }
     else {
-        delete_post_meta($post->ID, '_chi_selected_articles_or_videoss');
+        if (isset($post->ID)) {
+            delete_post_meta($post->ID, '_chi_selected_articles_or_videoss');
+        }
     }
 }
